requirements.txt
# Core dependencies
telethon==1.34.0
feedparser==6.0.10
praw==7.7.1
tweepy==4.14.0
requests==2.31.0
aiohttp==3.9.1
asyncio==3.4.3

# AI and NLP
langchain==0.1.0
langchain-community==0.0.10
ollama==0.1.7
spacy==3.7.2
nltk==3.8.1

# Database and caching
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Configuration and logging
python-dotenv==1.0.0
pydantic==2.5.2
loguru==0.7.2

# Data processing
pandas==2.1.4
beautifulsoup4==4.12.2
lxml==4.9.3

# Telegram Bot
python-telegram-bot==20.7

# Scheduling
schedule==1.2.0
apscheduler==3.10.4

# Utilities
pytz==2023.3
python-dateutil==2.8.2
